parametrize instantiation tests
